Consolidate examples into a single Xcode project
Following the convention of the CocoaPodsExampleLibrary[1], creates a
single Xcode project for both the table view and collection view
examples of SSDataSources.

An advantage of this strategy is a single Xcode project to hold the unit
tests for both the table view and collection view data source
responsibilities of SSDataSources.

[1]: https://github.com/CocoaPods/CocoaPodsExampleLibrary