Enable row editing by default
According to the documentation[1] in UITableView.h, if a
UITableViewDataSource object doesn't implement
-tableView:canEditRowAtIndexPath:, "all rows are assumed to be
editable".

Here, the core logic of SSBaseDataSource's implementation is to rely on
the fallbackTableDataSource if it implements the method, but otherwise
return a sane default. I think the sane default should mimic the
behavior of not implementing the method at all, so it should simply
return YES.

[1]: http://cl.ly/image/3i331s1I0G3e